// src/common/enums/taste-tag.ts
export enum TASTE_TAG {
  MAECONDALKOM = '매콤달콤한',
  DRINK_LIKE = '술_땡기는',
  LAST_SPICY = '끝맛이_매운',
  EOLEOL = '혀가_얼얼한',
  MARA = '마라맛',
  CAPSAICIN = '캡사이신맛',
  FIVE_STAR = '별이_다섯개',
  SOMETIME = '가끔_땡기는',
  SWEAT = '땀이_뻘뻘',
  TEARS = '눈물이_쏙',
  EOJIL = '어질어질한',
  MILK_ESSENTIAL = '우유가_필수',
  STRESS_RELIEF = '스트레스가_풀리는',
  JEONSULI_HOT = '정수리가_열나는'
}